refactor: change amount type

See merge request s06-blockchain-nft-sub2/S06P22A404!116

+++ doners-backend/src/main/java/com/doners/donersbackend/application/dto/request/donation/DonationInfoRequestDTO.java
@@ -56,7 +56,7 @@ public class DonationInfoRequestDTO {
 
     @ApiModelProperty(name = "목표 모금액")
     @Positive
-    private long targetAmount;
+    private double targetAmount;
 
 }
